gs-window-x11: shorten clock timeout to detect skew
The timer for the panel clock currently wakes up once a minute in
the default case.  This means, if the clock changes while the
screen is locked (say from a resume following suspend, or from
ntp) then it becomes out of date for up to a minute.

There's no good way right now to detect when clock skew happens,
so this commit changes the timeout to expire every couple of
seconds instead of once a minute.

Minor-Updates-By: Ray Strode <rstrode@redhat.com>

http://bugzilla.gnome.org/show_bug.cgi?id=648145