port gnome-screensaver to systemd-logind
On Fedora and other distributions ConsoleKit is being replaced by
systemd-logind. Accordingly, all tools previously talking to CK need to
talk to systemd-logind instead.

This patch will dynamically fall back to CK if systemd is not found,
thus being nice to Debian style distributions which want to support
systemd and non-systemd systems with the same packages.

https://bugzilla.gnome.org/show_bug.cgi?id=669787